[Edit] DrinkInfoView - 세그먼트 동작시 2개의 scrollView 각자의 offset 갖도록 수정 및 파일 분리 #14
